Start using NSInteger and NSUInteger
These have been introduced in Leopard and default to int and unsigned int.
In 64-bit Snow Leopard they are long and unsigned long.  This caused issues
with the getRectsBeingDrawn message which needs a pointer to a NSInteger
(long on 64-bit!) but we passed in an integer.  Surprisingly this problem
was visible when compiling with -O0 (segfault), but *not* when compiling
with -O1.  Other messages were NSInteger is now needed have also been
adapted.

Since NSInteger and NSUInteger are not available on Tiger, a define
has been added to add typedefs for these when they have not been defined
by the system headers.

(cherry picked from commit b2b70e5a8d275c8412709c9f077fb136078cd416)

+++ gdk/quartz/gdkevents-quartz.c
@@ -213,7 +213,7 @@ get_time_from_ns_event (NSEvent *event)
 static int
 get_mouse_button_from_ns_event (NSEvent *event)
 {
-  int button;
+  NSInteger button;
 
   button = [event buttonNumber];
